[libmicrohttpd] large file descriptors From: John Muth <muth@parascale.com>   To: libmicrohttpd@gnu.org   Date: Yesterday 18:35:25
Hi Christian,

We are using libmicro http in our application. Our application runs on Linux
and uses a large number of file descriptors (> 1024). Unfortunately the
Linux fdset's only have room for 1024 fd's.

I made changes to the 0.4.4 source to support poll instead of select so we
could get around the file descriptor issue. I've attached the diffs if you
are interested.

-John Muth
Parascale


git-svn-id: https://gnunet.org/svn/libmicrohttpd@9463 140774ce-b5e7-0310-ab8b-a85725594a96