[libmicrohttpd] Patches for Windows version of libmicrohttpd From: Gerrit Telkamp <g.telkamp@domologic.de>   To: libmicrohttpd@gnu.org   Date: Today 11:42:30   Attachments:  libmicrohttpd-windows.patch
  It was a bit tricky to build libmicrohttpd (SVN trunk) for Windows. 
I'm using the newest MSYS/MinGW environment (GCC 4.5.1).
I've optimized the autoconf feature, so this should be easier in future. 
Please find enclosed my patches.

In detail:
1- libmicrohttpd needs the "PlibC" library under Windows (libplibc.a). 
Methods of PlibC are referenced by src/daemon/daemon.c.
autoconf assumes that PlibC is installed on the Build environment and 
does not check this.
If PlibC is not installed, the pthreads check (see "acinclude.m4") will 
fail and stops with the wrong error message "Your system is not 
supporting pthreads". This is very confusing.

The patch includes an additional check for PlibC, before pthreads is 
used. The check is only done under cygwin and mingw environments.

Furthermore, an older version of plibc.h was in the include directory. 
This gives some problems if you are using a newer version of the PlibC 
library. So I propose to exclude src/include/plibc.h from the SVN

Files affected:
acinclude.m4 (modified)
configure.ac (modified)
src/include/plibc.h (deleted)

2- setsockopt() uses a different parameter type under Windows ("optval" 
is const char). This caused an error when src/daemon/daemon.c was compiled.

The patch enclosed includes a different call of the setsockopt() on 
Windows machines.
Important: IPV6_V6ONLY is not available on Windows machines < Windows 7! 
So this parameter is "0". But there is a TODO comment, maybe someone has 
an idea how to solve this on Windows XP machines.

Files affected:
src/daemon/daemon.c (modified)

I have tested the patch on a MinGW / MSYS environment only.

Best regards

Gerrit.

libmicrohttpd-windows.patch


git-svn-id: https://gnunet.org/svn/libmicrohttpd@12672 140774ce-b5e7-0310-ab8b-a85725594a96